requirements.txt
# FastAPI and ASGI server (версии с готовыми бинарниками под Windows)
fastapi>=0.121.1,<0.122
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.23
asyncpg>=0.29.0
psycopg[binary]>=3.1.14
alembic>=1.12.1

# Authentication
PyJWT[crypto]>=2.10.1
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Validation (только бинарные сборки — без Rust)
pydantic>=2.5.0,<3
pydantic-settings>=2.1.0

# Utilities
python-dateutil>=2.8.2
openpyxl>=3.1.0
aiofiles>=23.2.1

# Logging
loguru>=0.7.2
sentry-sdk>=2.0.0

# Telegram bot
python-telegram-bot>=21.11,<22
httpx>=0.28.1,<0.29

# Тестирование интерфейса
playwright>=1.40.0

# Security floor (CVE fixes observed by safety scan)
requests>=2.32.5
urllib3>=2.6.3
starlette>=0.50.0,<0.51
jinja2>=3.1.6
pillow>=12.1.1
fonttools>=4.61.1
pyasn1>=0.6.2
peewee>=4.0.0
